requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-dht==4.0.5
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorzero==2.0
Flask==3.0.3
Flask-SocketIO==5.4.1
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gpiozero==2.0.1
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==3.2.2
paho-mqtt==2.1.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyftdi==0.55.4
pyparsing==3.2.0
pyserial==3.5
python-engineio==4.10.1
python-socketio==5.11.4
pyusb==1.2.1
requests==2.32.3
requests-oauthlib==2.0.0
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
rsa==4.9
simple-websocket==1.1.0
SQLAlchemy==2.0.36
sysv_ipc==1.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.6
wsproto==1.2.0
